cwl: try to include validate plugin
cwl-svg offers a validation plugin that we should be able to use to
validate workflows on the fly, as it will highlight edges that are
invalid (from looking in the source code). however I couldn't get this to work - maybe it is impossible to create an invalid workflow from the GUI and this only functions for workflows created from a text editor?

Signed-off-by: Jacob <jlh2hb@virginia.edu>